Added utils file. Added rtl support. Refactored logic file
